+++ zimui/src/router/index.ts
@@ -4,7 +4,10 @@ import HomeView from '../views/HomeView.vue'
 import VideoPlayerView from '../views/VideoPlayerView.vue'
 import NotFoundView from '../views/NotFoundView.vue'
 
+import ChannelHomeTab from '@/components/channel/tabs/ChannelHomeTab.vue'
 import VideosTab from '@/components/channel/tabs/VideosTab.vue'
+import ShortsTab from '@/components/channel/tabs/ShortsTab.vue'
+import LivesTab from '@/components/channel/tabs/LivesTab.vue'
 import PlaylistsTab from '@/components/channel/tabs/PlaylistsTab.vue'
 import PlaylistView from '@/views/PlaylistView.vue'
 
@@ -15,12 +18,27 @@ const router = createRouter({
       path: '/',
       name: 'home',
       component: HomeView,
-      redirect: '/videos',
+      redirect: '/channel-home',
       children: [
+        {
+          path: 'channel-home',
+          name: 'channel-home',
+          component: ChannelHomeTab
+        },
         {
           path: 'videos',
           name: 'videos',
           component: VideosTab
+        },
+        {
+          path: 'shorts',
+          name: 'shorts',
+          component: ShortsTab
+        },
+        {
+          path: 'lives',
+          name: 'lives',
+          component: LivesTab
         },
         {
           path: 'playlists',